renamed the Tags folder and the files inside it
